+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+  - Voucher - Decommission voucher resources
+
 ## [5.9.0] - 2021-02-16
 
 ### Changed